refactor: ProfitRate 포장
