+++ src/main/java/lotto/domain/summary/Summary.java
@@ -1,8 +1,5 @@
 package lotto.domain.summary;
 
-import lotto.constants.Winning;
-
-import java.util.Map;
 import java.util.Objects;
 
 public class Summary {
@@ -10,26 +7,11 @@ public class Summary {
     public static final String PROFIT_RATE_MESSAGE = "총 수익률은 %f입니다.";
 
     private final WinningSummary winningSummary;
-    private final float profitRate;
+    private ProfitRate profitRate;
 
     public Summary(WinningSummary winningSummary, Long purchasePrice) {
         this.winningSummary = winningSummary;
-        this.profitRate = profitRate(winningSummary.winnings(), purchasePrice);
-    }
-
-
-    private float profitRate(Map<Winning, Long> winnings, long purchasePrice) {
-        return (float) prizeTotal(winnings) / (float) purchasePrice;
-    }
-
-    private long prizeTotal(Map<Winning, Long> winnings) {
-        return winnings.entrySet().stream()
-                .mapToLong(this::prizeTotalByWinning)
-                .sum();
-    }
-
-    private long prizeTotalByWinning(Map.Entry<Winning, Long> winning) {
-        return winning.getKey().prize() * winning.getValue();
+        this.profitRate = new ProfitRate(winningSummary, purchasePrice);
     }
 
     @Override
@@ -37,7 +19,7 @@ public class Summary {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Summary summary = (Summary) o;
-        return Float.compare(profitRate, summary.profitRate) == 0 && Objects.equals(winningSummary, summary.winningSummary);
+        return Objects.equals(winningSummary, summary.winningSummary) && Objects.equals(profitRate, summary.profitRate);
     }
 
     @Override
@@ -49,7 +31,7 @@ public class Summary {
     public String toString() {
         StringBuilder stringBuffer = new StringBuilder();
         stringBuffer.append(winningSummary);
-        stringBuffer.append(String.format(PROFIT_RATE_MESSAGE, profitRate));
+        stringBuffer.append(String.format(PROFIT_RATE_MESSAGE, profitRate.value()));
 
         return stringBuffer.toString();
     }
